# CUDA: enable peer access between devices