This PR enables peer access between CUDA devices if possible. As a consequence devices can communicate directly via PCIe instead of using the CPU as an intermediary. This makes token generation faster:

 | GPU    | Model    | Test  | t/s 1x P40 | t/s master | t/s PR | Speedup |
|--------|----------|-------|------------|------------|--------|---------|
| 3x P40 | 7b q4_0  | tg128 |      50.29 |      44.48 |  56.23 |    1.26 |
| 3x P40 | 13b q4_0 | tg128 |      27.86 |      30.32 |  37.46 |    1.24 |
| 3x P40 | 33b q4_0 | tg128 |      12.11 |      15.85 |  18.54 |    1.17 |
| 3x P40 | 70b q6_K | tg128 |          - |       7.36 |   8.15 |    1.11 |
| 3x P40 | 7b q4_0  | pp    |     703.32 |     384.37 | 365.01 |    0.95 |
| 3x P40 | 13b q4_0 | pp    |     384.52 |     243.87 | 224.95 |    0.92 |
| 3x P40 | 33b q4_0 | pp    |     158.15 |     121.24 | 111.78 |    0.92 |
| 3x P40 | 70b q6_K | pp    |          - |      58.15 |  54.86 |    0.94 |

However, for some reason that I don't yet understand it also makes prompt processing slightly slower. Peer access makes memory allocation slower but I don't think that is the cause. In any case, I think even with the decrease in prompt processing speed this would be a net positive.