@JohannesGaessler 2023-07-31
I forgot to add: this PR would also enable NVLink if someone were to use it (P40s do not have NVLink).