@slaren 2023-07-31
I tried this with 3090 Ti + 3080 on WSL2 and there isn't really a difference, however I am limited to 7B because it fails with "out of memory" errors for anything bigger than that. I suspect that this is because I cannot enable above 4G decoding with both GPUs on my system. The performance with 7B and the two GPUs is very bad for me, from 25 ms/token with a single GPU, to 170 ms/token with two GPUs.

I wonder if writing the result directly to memory of the main GPU in the kernel would be faster than the memcpy.