@JohannesGaessler 2023-07-31
Direct communication without NVLink via PCIe only works on Linux I think. You should be able to check this via `nvidia-smi topo -p2p r`.