@slaren 2023-07-31
Yes you're right, `cudaDeviceCanAccessPeer` returns 0. `nvidia-smi` says `CNS  = Chipset not supported`, which again seems related to the above 4G decoding issue.