@JohannesGaessler 2023-07-31
Some profiling data for perplexity calculations using 7b q4_0:

| CUDA API call     | Total no peer access | Per call no peer access | Per call with peer access | Total with peer access |
|-------------------|----------------------|-------------------------|---------------------------|------------------------|
| cudaMemcpyAsync   | 11.798 s             | 2.362 ms                | 8.322 µs                  | 41.563 ms              |
| cudaLaunchKernel  | 546.796 ms           | 8.340 µs                | 114.965 µs                | 7.537 s                |
| cudaMemcpy2DAsync | 261.229 ms           | 9.503 µs                | 190.392 µs                | 5.234 s                |

It seems that enabling peer access makes `cudaMemcpyAsync` much faster but it also makes `cudaMemcpy2DAsync` slower and greatly increases kernel launch overhead.