@Ph0rk0z 2023-08-22
On 2x3090 it improved top t/s beyond exllama. I have actual nvlink. 13t/s in textgen vs 10t/s. For prompt processing, it stopped being noticeable on at least mid length contexts (1-2k). Time to first token was making me want to go back to GPTQ.

Did not try to split a model across all 3 of my cards yet post change. I've got the same exact xeon v4 as you but all layers are offloaded.