@JohannesGaessler 2023-08-27
I currently don't have GGUF models higher than 7b ready to re-test performance but for 7b the relative performance after rebasing seems to be the same.