@Ph0rk0z 2023-08-27
Was waiting on llama python bindings to catch up to test in the new implementation.

When I ran 70b Q6 I added a string that peering was enabled. It enabled for the 3090s but not the P40. Still, speed for a 3 card model of 2 different generations isn't bad. This cannot be done with GPTQ without 3x3090.