@pancoaster 2023-08-30
Greetings, from my understanding of this thread, the implementation slows down prompt processing but does in fact enable loading larger models into pooled VRAM. IMO this is a valuable feature, maybe enable merge depending on a CMAKE flag?