@JohannesGaessler 2023-08-30
As I said in the issue: the proper way to implement this would be to not just enable peer access but to then also adapt `ggml_cuda_op`. If this also fixes the prompt processing speed we don't need to add a compile option and can just enable peer access unconditionally if it is supported. I want to do this at some point but since you said "Willing and able to implement anything required to enabled pooled VRAM. Looking for direction and suggested starting points to generate the PR." I would also be fine with letting you take a crack at it.