@Ph0rk0z 2023-08-30
I do not get any kind of processing slowdown with proper nvlink. It just beats exllama, simple as. There is no tradeoff for me with fully offloaded models that I can tell. I rather get 17t/s than 10t/s. If there is a slowdown it will happen to 220+ t/s prompt processing and I'll take it.

I am really unsure why people think there is pooled memory. Even with Nvlink it's not pooled. Simply the cards skip the step of sending anything back to the CPU and communicate directly.