@pancoaster 2023-08-30
I’m noobing here. Explain how to load the 70B model with full cuda offload to two 3090ti. What hardware is required to upgrade from a system with one 3090ti to match your results? Nvlink not needed?