@Ph0rk0z 2023-08-30
I just tell it to offload 83 layers and then do the memory split like 41,42

You don't need nvlink but you'll get the 10 it/s rather than more. Definitely need 48GB and more for quants above 4KM. The highest I ran is Q6 but that requires using the P40 and then I only get 7 it/s