@pancoaster 2023-08-31
What would be valuable additions if I request access to a100? How can I write a proposal for access based on this repo? Curious if it’s worth my time to try to get access to the big iron cluster or just add another card to my personal system