@Ph0rk0z 2023-08-31
Which A100? If it's the 40gb you will come up short.