@pancoaster 2023-08-31
Hoping to program Ascend for something entertaining: https://www.osc.edu/resources/technical_support/supercomputers/ascend