@Ph0rk0z 2023-08-31
You can train quite well there. I would leave inference to 2x3090 setups.Training 70b on only 48g is slow, even as Q4.