@pancoaster 2023-08-31
So would you recommend the proposal output to be new model weights after training the base 70B model with more data? Trying to envision why a project about llama.cpp would get funded.