@Ph0rk0z 2023-08-31
I don't know what proposal you could write. If you had an idea for a dataset to produce some kind of model I'm sure they would give you compute time. But just to run inference I think you are better off buying another card.

Off the top of my head I would combine teatime+todd proxy outputs with first cleaned orca, guanaco and then airoboros to see which one would work better. But I doubt someone would fund my pet anti-alignment projects.