@JohannesGaessler 2023-08-31
So I tried an implementation where GPUs directly read to/write from the VRAM of other GPUs but as it turns out that is slower than copying in one batch at the beginning or the end. So I think the way to go is to add a toggle (with enabled peer access being the default). 

More generally, it seems that the biggest source of kernel launch overhead is the cuBLAS GEMM for the KV cache (due to the many attention heads). For token generation the overhead is much lower because the kernels I wrote can process all attention heads at once and I think that's why enabling peer access makes it faster.