@Ph0rk0z 2023-09-01
You did something because exllama also enables peer access and it's slower.