@JohannesGaessler 2023-09-10
After https://github.com/ggerganov/llama.cpp/pull/3110 I finally understand the prompt processing performance regression from enabling peer access. The problem is that by default the data goes device -> host -> device with buffering on the host. If you enable peer access the data goes directly device -> device with no buffering so for 2+ GPUs communicating with the main device they can block each others' data transfer and the performance goes down.

Unfortunately you apparently cannot just toggle peer to peer data transfer on a per-API-call basis though. The proper way to fix this would be to reduce the bandwidth needed for data transfers (e.g. by transferring f16 or q8_0) or by manually implementing a buffered device -> host -> device data transfer for the writeback to the main GPU. But for now I think we can add a function that sets peer access depending on batch size and NVLink availability and call it at the beginning of each eval.