@JohannesGaessler 2023-09-10
>The proper way to fix this would be to reduce the bandwidth needed for data transfers (e.g. by transferring f16 or q8_0) or by manually implementing a buffered device -> host -> device data transfer for the writeback to the main GPU.

Actually, now that I think about it it may be possible to fix the performance by reordering the data transfers via CUDA events.