@Ph0rk0z 2023-09-10
Won't this do nothing as is because `#ifdef NDEBUG`

I took this out when testing.