@JohannesGaessler 2023-09-10
`NDEBUG` is defined unless you enable debugging, so by default it _should_ do something.