@JohannesGaessler 2023-09-16
I tried an implementation where I used CUDA events to enforce a specific data transfer order but the performance was still worse than with peer access enabled. I now did an implementation that at the beginning of an eval enables or disables peer access based on batch size. The threshold for peer access can be controlled via a compile option `LLAMA_CUDA_PEER_MAX_BATCH_SIZE` and the default is 128. I am basing this value on the following results that I got with `LLAMA_CUDA_PEER_MAX_BATCH_SIZE=512`:

| model                          | test       |    n_batch | t/s master       | t/s PR           | Speedup |
| ------------------------------ | ---------- | ---------: | ---------------: | ---------------: | ------- |
| LLaMA 7B mostly Q4_0           | pp 512     |          1 | 49.21 ± 0.04     | 67.04 ± 0.11     |    1.36 |
| LLaMA 7B mostly Q4_0           | pp 512     |          2 | 31.71 ± 0.01     | 35.55 ± 0.01     |    1.12 |
| LLaMA 7B mostly Q4_0           | pp 512     |          4 | 61.18 ± 0.01     | 68.49 ± 0.02     |    1.12 |
| LLaMA 7B mostly Q4_0           | pp 512     |          8 | 108.74 ± 0.02    | 119.72 ± 0.04    |    1.10 |
| LLaMA 7B mostly Q4_0           | pp 512     |         16 | 193.34 ± 0.08    | 214.39 ± 0.09    |    1.11 |
| LLaMA 7B mostly Q4_0           | pp 512     |         32 | 321.10 ± 0.06    | 354.31 ± 0.22    |    1.10 |
| LLaMA 7B mostly Q4_0           | pp 512     |         64 | 478.07 ± 0.16    | 524.90 ± 0.23    |    1.10 |
| LLaMA 7B mostly Q4_0           | pp 512     |        128 | 588.65 ± 0.21    | 638.28 ± 0.28    |    1.08 |
| LLaMA 7B mostly Q4_0           | pp 512     |        256 | 744.01 ± 0.43    | 595.75 ± 0.06    |    0.80 |
| LLaMA 7B mostly Q4_0           | pp 512     |        512 | 904.99 ± 0.37    | 708.86 ± 0.62    |    0.78 |
| LLaMA 7B mostly Q4_0           | tg 128     |          1 | 52.95 ± 0.02     | 71.53 ± 0.07     |    1.35 |

With NVLink enabled a higher value would probably perform better but unfortunately CUDA does not seem to let you query NVLink status between devices.